json parse with atoms as keys, copy_doc
Parse all db responses into proper maps with atoms as keys. we will not
be overrun with keys as long as the db documents dont have a limited
number of keys. This is probably a bad idea, opening us up for atom
overrun attacks. We could use the ```:atoms!``` option for the parser,
but then it would be too restricted in most cases. Its probably best to
make it an option with a reasonablle default setting.

In other news: copy_doc